Update README.md
A friend said it is best to demo apps during interviews and haveing everything ready deployed for easy access AKA heroku but Molly said to better have it linked from github so I decide to add a couple of lines to the readme.  And also brag a little about the technologies we used.
Please feel free to correct me if I describe the technology wrong or if you guys feel its better to leave out the technology we used.